Move app action buttons to detail screen
In some language, screen size, display size setting, and font size
setting combinations, the install/uninstall/update/open buttons in app
cards are space constrained, causing the buttons to wrap their text or
even disappear completely.

Fix this by moving the buttons to the app detail screen which has plenty
of space to spare.

Fixes #129
